+++ src/helpers/code-lens-actions.helper.ts
@@ -1,0 +1,17 @@
+import { MatchTypeChangeAnnotations } from "./types";
+import * as vscode from "vscode";
+
+export function changeTestAnnotations(match: MatchTypeChangeAnnotations): void {
+  const textEditor = vscode.window.activeTextEditor;
+  if (!textEditor) {
+    vscode.window.showErrorMessage("Editor Does Not Exist");
+    return;
+  }
+
+  let textFragment = textEditor.document.getText(match.range);
+  const newFragment = textFragment.replace(match.from, match.to);
+
+  textEditor.edit((editBuilder) => {
+    editBuilder.replace(match.range, newFragment);
+  });
+}